requirements.txt
shiny
pandas
jinja2
matplotlib
seaborn
python-dotenv